--- Frontend/yummy_market/src/components/login/Login.tsx
@@ -1,10 +1,0 @@
-export default function Login()
-{
-  return(
-      <>
-        <div>
-          <input type="text"/>
-        </div>
-      </>
-  );
-}
