+++ Frontend/yummy_market/src/styles/login.css
@@ -67,3 +67,55 @@ padding: 50px 0;
   background: #fff;
   color: #111;
 }
+
+
+/* login.css 파일에 추가 */
+.tabs-container {
+  display: flex;
+  margin-bottom: 20px;
+  border-bottom: 1px solid #ddd;
+}
+
+.tab-button {
+  padding: 12px 0;
+  background: none;
+  border: none;
+  cursor: pointer;
+  transition: all 0.3s ease;
+  color: #888;
+  position: relative;
+}
+
+.tab-button.active-tab {
+  color: #000;
+  font-weight: 500;
+}
+
+.tab-button.active-tab::after {
+  content: '';
+  position: absolute;
+  bottom: -1px;
+  left: 0;
+  width: 100%;
+  height: 2px;
+  background-color: #000;
+}
+
+.tab-button:hover {
+  color: #000;
+}
+
+.verification-btn {
+  position: absolute;
+  right: 10px;
+  background-color: #f0f0f0;
+  border: none;
+  border-radius: 4px;
+  padding: 5px 10px;
+  cursor: not-allowed;
+}
+
+.opacity-50 {
+  opacity: 0.5;
+  cursor: not-allowed;
+}
